Add location context and update settings screen
