+++ app/src/Context/GlobalState.tsx
@@ -9,13 +9,12 @@ import {
   CardBank,
   CardBrand,
 } from '../types/CardType';
-import {AppState, Keyboard, PermissionsAndroid, Platform} from 'react-native';
-import Geolocation from 'react-native-geolocation-service';
+import {AppState, Keyboard} from 'react-native';
 import Consts from '../Helpers/Consts';
 import Config from 'react-native-config';
 import {AuthContextType} from '../types/AuthContextType';
 import AuthContext from './authContext';
-import {Location, Place, PlaceLocation} from '../types/Location';
+import LocationState from './LocationState';
 const baseURL = Config.REACT_APP_API_URL;
 
 const GlobalState: React.FC<PropsWithChildren> = ({children}) => {
@@ -23,10 +22,6 @@ const GlobalState: React.FC<PropsWithChildren> = ({children}) => {
     React.useContext(AuthContext) as AuthContextType;
 
   const [isLoading, setIsLoading] = React.useState<boolean>(true);
-  const [location, setLocation] = useState<Location>({} as Location);
-  const [places, setPlaces] = useState<Place[]>([] as Place[]);
-  const [address, setAddress] = useState<Place | undefined>(undefined);
-  const [locationGrantType, setLocationGrantType] = useState<string>('denied');
 
   const [rewards] = React.useState<Reward[]>(Consts.dummyCardRewards);
   const ErrorMessages = Consts.authErrorMessages;
@@ -308,303 +303,6 @@ const GlobalState: React.FC<PropsWithChildren> = ({children}) => {
     return errors;
   }
 
-  const requestLocationPermission = useCallback(async () => {
-    try {
-      if (Platform.OS === 'ios') {
-        const grant = await Geolocation.requestAuthorization('whenInUse');
-        setLocationGrantType(grant);
-      } else {
-        const grant = await PermissionsAndroid.request(
-          PermissionsAndroid.PERMISSIONS.ACCESS_FINE_LOCATION,
-          {
-            title: 'Geolocation Permission',
-            message: 'Can we access your location?',
-            buttonNeutral: 'Ask Me Later',
-            buttonNegative: 'Cancel',
-            buttonPositive: 'OK',
-          },
-        );
-        setLocationGrantType(grant);
-      }
-      if (locationGrantType === 'denied') {
-        return false;
-      } else {
-        return true;
-      }
-    } catch (err) {
-      return false;
-    }
-  }, [locationGrantType]);
-
-  const fetchAddress = useCallback(async () => {
-    const permission = await requestLocationPermission();
-    if (permission === false) {
-      setAddress({
-        businessStatus: 'OPERATIONAL',
-        displayName: {text: 'Location Permission Denied', languageCode: ''},
-        location: {longitude: 0, latitude: 0},
-        primaryType: 'restaurant',
-        primaryTypeDisplayName: {text: 'Restaurant', languageCode: ''},
-        types: ['Restaurant'],
-        readableType: 'Restaurant',
-        id: '0',
-      });
-      return;
-    }
-    var myHeaders = new Headers();
-    myHeaders.append('Content-Type', 'application/json');
-    myHeaders.append(
-      'X-Goog-FieldMask',
-      'places.displayName,places.businessStatus,places.primaryType',
-    );
-    myHeaders.append(
-      'X-Goog-Api-Key',
-      'AIzaSyDSQqzE6cXDeUCWEquYC4PPCCpk9KRJiw8',
-    );
-
-    var raw = JSON.stringify({
-      excludedTypes: ['parking'],
-      maxResultCount: 1,
-      rankPreference: 'DISTANCE',
-      locationRestriction: {
-        circle: {
-          center: {
-            latitude: location.latitude,
-            longitude: location.longitude,
-          },
-          radius: 50000,
-        },
-      },
-    });
-
-    const response = await fetch(
-      'https://places.googleapis.com/v1/places:searchNearby',
-      {
-        method: 'POST',
-        headers: myHeaders,
-        body: raw,
-      },
-    );
-
-    // Manipulate result to return
-    const result = await response.json();
-    const resultAddress = result.places.map((place: Place, index: number) => {
-      return {
-        ...place,
-        id: index.toString(),
-      } as Place;
-    });
-
-    setAddress(resultAddress[0]);
-  }, [location, requestLocationPermission]);
-
-  const calculateDistanceLatLong = (
-    location1: PlaceLocation,
-    location2: PlaceLocation,
-  ) => {
-    const toRadians = (degrees: number): number => {
-      return degrees * (Math.PI / 180);
-    };
-
-    const earthRadius = 3958.8;
-    const lat1 = location1.latitude;
-    const lat2 = location2.latitude;
-    const lon1 = location1.longitude;
-    const lon2 = location2.longitude;
-    const sinLat1 = Math.sin(toRadians(lat1));
-    const sinLat2 = Math.sin(toRadians(lat2));
-    const cosLat1 = Math.cos(toRadians(lat1));
-    const cosLat2 = Math.cos(toRadians(lat2));
-    const lonDifference = toRadians(lon2 - lon1);
-    const distance =
-      Math.acos(
-        sinLat1 * sinLat2 + cosLat1 * cosLat2 * Math.cos(lonDifference),
-      ) * earthRadius;
-    return Math.round(distance * 100) / 100;
-  };
-
-  const fetchPlaces = useCallback(async () => {
-    const permission = await requestLocationPermission();
-    if (permission === false) {
-      setPlaces([
-        {
-          businessStatus: 'OPERATIONAL',
-          displayName: {text: 'Location Permission Denied', languageCode: ''},
-          location: {longitude: 0, latitude: 0},
-          primaryType: 'restaurant',
-          primaryTypeDisplayName: {text: 'Restaurant', languageCode: ''},
-          types: ['Restaurant'],
-          readableType: 'Restaurant',
-          id: '0',
-        },
-      ]);
-      return;
-    }
-    var headers = new Headers();
-    headers.append('Content-Type', 'application/json');
-    headers.append(
-      'X-Goog-FieldMask',
-      'places.displayName,places.businessStatus,places.primaryType,places.location,places.primaryTypeDisplayName,places.types',
-    );
-    headers.append('X-Goog-Api-Key', 'AIzaSyDSQqzE6cXDeUCWEquYC4PPCCpk9KRJiw8');
-
-    var placesTypes = {
-      restaurant: 'Restaurant',
-      museum: 'Museum',
-      movie_theater: 'Movie Theater',
-      gas_station: 'Gas Station',
-      car_wash: 'Car Wash',
-      car_repair: 'Car Repair',
-      car_rental: 'Car Rental',
-      car_dealer: 'Car Dealer',
-      electric_vehicle_charging_station: 'EV Charging Station',
-      rest_stop: 'Rest Stop',
-    };
-
-    var raw = JSON.stringify({
-      includedTypes: Object.keys(placesTypes),
-      maxResultCount: 20,
-      rankPreference: 'DISTANCE',
-      locationRestriction: {
-        circle: {
-          center: {
-            latitude: location.latitude,
-            longitude: location.longitude,
-          },
-          radius: 50000,
-        },
-      },
-    });
-
-    const response = await fetch(
-      'https://places.googleapis.com/v1/places:searchNearby',
-      {
-        method: 'POST',
-        headers: headers,
-        body: raw,
-      },
-    );
-
-    // Manipulate result to return
-    const result = await response.json();
-    const resultPlaces = result.places
-      .filter((place: Place) => {
-        return place.businessStatus === 'OPERATIONAL';
-      })
-      .map((place: Place, index: number) => {
-        let primaryTypeDisplayName = {};
-        if (
-          !place.hasOwnProperty('primaryType') &&
-          place.types.length > 0 &&
-          placesTypes.hasOwnProperty(place.types[0].toString())
-        ) {
-          primaryTypeDisplayName = {
-            lang: 'en-US',
-            text: placesTypes[place.types[0] as keyof typeof placesTypes],
-          };
-        } else if (!place.hasOwnProperty('primaryType')) {
-          let type = place.types[0];
-          let displayName = '';
-          type.split('_').forEach(name => {
-            displayName +=
-              name.charAt(0).toUpperCase() + name.substring(1) + ' ';
-          });
-          primaryTypeDisplayName = {
-            lang: 'en-US',
-            text: displayName,
-          };
-        } else {
-          primaryTypeDisplayName = place.primaryTypeDisplayName;
-        }
-        return {
-          ...place,
-          primaryTypeDisplayName: primaryTypeDisplayName,
-          distance: calculateDistanceLatLong(location, place.location),
-          id: index.toString(),
-        } as Place;
-      });
-
-    setPlaces(resultPlaces);
-  }, [location, requestLocationPermission]);
-
-  const getLocation = useCallback(() => {
-    const result = requestLocationPermission();
-    result.then(res => {
-      if (res) {
-        Geolocation.getCurrentPosition(
-          position => {
-            const coords = position.coords;
-            console.log(coords);
-            setLocation({
-              latitude: coords.latitude,
-              longitude: coords.longitude,
-              altitude: coords.altitude,
-              accuracy: coords.accuracy,
-            } as Location);
-          },
-          error => {
-            // See error code charts below.
-            setLocation({} as Location);
-            console.log(error.code, error.message);
-          },
-          {enableHighAccuracy: true, timeout: 15000, maximumAge: 1},
-        );
-      } else {
-        //Basic location
-        setLocation({
-          latitude: 0,
-          longitude: 0,
-          altitude: 0,
-          accuracy: 0,
-        } as Location);
-      }
-    });
-  }, [requestLocationPermission]);
-
-  const updateLocation = useCallback(() => {
-    setIsLoading(true);
-    const result = requestLocationPermission();
-    result.then(res => {
-      if (res) {
-        Geolocation.getCurrentPosition(
-          position => {
-            const coords = position.coords;
-            setLocation({
-              latitude: coords.latitude,
-              longitude: coords.longitude,
-              altitude: coords.altitude,
-              accuracy: coords.accuracy,
-            } as Location);
-          },
-          error => {
-            // See error code charts below.
-            setLocation({
-              latitude: 0,
-              longitude: 0,
-              altitude: 0,
-              accuracy: 0,
-            } as Location);
-            console.log(error.code, error.message);
-          },
-          {enableHighAccuracy: true, timeout: 15000, maximumAge: 1},
-        );
-      } else {
-        //Basic location
-        setLocation({
-          latitude: 0,
-          longitude: 0,
-          altitude: 0,
-          accuracy: 0,
-        } as Location);
-      }
-    });
-    setIsLoading(false);
-  }, [requestLocationPermission]);
-
-  useEffect(() => {
-    getLocation();
-  }, [getLocation]);
-
   const fetchBanks = useCallback(async () => {
     const headers = new Headers();
     headers.append('Content-Type', 'application/json');
@@ -652,11 +350,6 @@ const GlobalState: React.FC<PropsWithChildren> = ({children}) => {
     }
   }, [fetchBanks, fetchBrands, userToken]);
 
-  useEffect(() => {
-    fetchPlaces();
-    fetchAddress();
-  }, [fetchAddress, fetchPlaces, location]);
-
   const [isKeyboardVisible, setKeyboardVisible] = useState(false);
 
   useEffect(() => {
@@ -702,7 +395,6 @@ const GlobalState: React.FC<PropsWithChildren> = ({children}) => {
         unlinkCard,
         addNewReward,
         removeReward,
-        location,
         isLoading,
         setIsLoading,
         bankOptions,
@@ -713,16 +405,10 @@ const GlobalState: React.FC<PropsWithChildren> = ({children}) => {
         validateCardForm,
         newCardBin,
         setNewCardBin,
-        fetchPlaces,
-        places,
-        fetchAddress,
-        address,
         isKeyboardVisible,
-        requestLocationPermission,
         appStateVisible,
-        updateLocation,
       }}>
-      {children}
+      <LocationState>{children}</LocationState>
     </Context.Provider>
   );
 };
