+++ app/src/Context/locationContext.tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import {LocationContext} from '../types/LocationContextType';
+
+export default React.createContext<LocationContext | null>(null);
